src/benchmarks/vector-rag/pyproject.toml
[project]
name = "vector-rag-bench"
version = "0.1.0"
description = "Vector RAG bench (LangChain + FAISS + OpenAI-compatible API)"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.8.0",
]
